added answers.js action, changed saveAnswer -> updateAnswer
