+++ app/javascript/packs/actions/buildings.js
@@ -18,10 +18,6 @@ import {
   SAVE_ANSWER,
 } from '../constants';
 
-  /////////////////////////////////////////////////////////////////
- // BUILDINGS ////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////
-
 export function addBuilding(portfolioId) {
   return {
     type: ADD_BUILDING,
@@ -113,59 +109,3 @@ export async function updateBuilding(id, updatedBuilding, dispatch) {
   };
 }
 
-  /////////////////////////////////////////////////////////////////
- // ANSWERS //////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////
-
-export function createAnswer(buildingId) {
-  return {
-    type: CREATE_ANSWER,
-    status: FETCH_IN_PROGRESS,
-    buildingId: buildingId,
-    questionId: questionId
-  };
-}
-
-// export function updateAnswer() {
-//   return {
-//     type: UPDATE_ANSWER,
-//     portfolioId
-//   };
-// }
-
-function saveAnswer(type, response) {
-  return {
-    type,
-    status: FETCH_SUCCESS,
-    response
-  };
-}
-
-function saveError(type, error) {
-  return {
-    type,
-    status: FETCH_FAILURE,
-    response: error
-  };
-}
-
-
-export async function addAnswer(buildingId, answer, dispatch) {
-  dispatch(createAnswer(buildingId));
-
-  try {
-    const answerData = JSON.stringify({
-      ...answer
-    });
-
-    let response = await fetch('/api/buildings', {
-      method: 'POST',
-      data: answerData
-    }).then(resp => resp.json());
-
-    dispatch(updateAnswer());
-    dispatch(saveAnswer(ADD_ANSWER, response.data));
-  } catch (error) {
-    dispatch(saveError(CREATE_ANSWER, error));
-  }
-}
